requirements.txt
pydantic>=2.5
ortools>=9.8
numpy
pytest>=9.0.2
fastapi>=0.110
uvicorn>=0.29
openpyxl>=3.1.5
scikit-learn>=1.8.0
requests>=2.31.0
folium>=0.15.0
pandas
python-multipart>=0.0.6
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
